split methods supporting different QC plots into different files
