+++ src/ophys_etl/modules/decrosstalk/qc_plotting/__init__.py
@@ -1,0 +1,5 @@
+from .summary_plot import generate_roi_figure
+from .pairwise_plot import generate_pairwise_figures
+
+
+__all__ = ["generate_roi_figure", "generate_pairwise_figures"]
